requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
requests==2.32.3
python-dateutil==2.9.0.post0
google-cloud-vision==3.7.2
google-auth==2.29.0
tzdata



